EatV2/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
mss>=9.0.1
pyautogui>=0.9.54
pynput>=1.7.6
Pillow>=10.0.0
